% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avonet-package.R
\docType{package}
\name{avonet-package}
\alias{avonet}
\alias{avonet-package}
\title{avonet: Morphological, Ecological and Geographical Data for All Birds}
\description{
An R interface designed to access the AVONET database of avian functional traits by providing easy applicable functions for the basic steps of data preprocessing, e.g. data import & exploration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kfrac/avonet}
  \item Report bugs at \url{https://github.com/kfrac/avonet/issues}
}

}
\author{
\strong{Maintainer}: Kevin Frac \email{kevin.frac@senckenberg.de} (\href{https://orcid.org/0009-0004-8104-8627}{ORCID})

}
\keyword{internal}
