% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to AVONET database}
\usage{
connect(username, pw)
}
\arguments{
\item{username}{Username for AVONET database}

\item{pw}{Password for AVONET database}
}
\value{
A database connection
}
\description{
Connect to AVONET database
}
\examples{
con <- connect(user = "postgres", password = "Frankfurterstr25!")
}
