% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{disconnect_db}
\alias{disconnect_db}
\title{Disconnect from AVONET database}
\usage{
disconnect_db()
}
\value{
Invisibly, \code{NULL}.
}
\description{
Closes the active database connection opened by \code{\link[=connect_db]{connect_db()}}, if there
is one.
}
\examples{
\dontrun{
db_user <- keyring::key_list("avonet")[1, "username"]
db_password <- keyring::key_get("avonet", username = db_user)
connect_db(username = db_user, pw = db_password)

# ...do some work...

disconnect_db()
}
}
\seealso{
\code{\link[=connect_db]{connect_db()}}
}
