% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomic_info.R
\name{get_taxonomic_info}
\alias{get_taxonomic_info}
\title{Retrieve taxonomic information}
\usage{
get_taxonomic_info(search_term, taxonomy)
}
\arguments{
\item{search_term}{Name of a genus, family or order}

\item{taxonomy}{Choose which taxonomy your results are displayed in. 1 = BirdLife, 2 = eBird and 3 = BirdTree}
}
\value{
A dataframe
}
\description{
Retrieve taxonomic information
}
\examples{
db_user <- keyring::key_list("avonet")[1, "username"]
db_password <- keyring::key_get("avonet", username = db_user)
connect_db(username = db_user, pw = db_password)

# This works for a single species
get_taxonomic_info(search_term = "Buteo buteo", taxonomy = 1)

# Or a list of several species
get_taxonomic_info(search_term = "passeriformes", taxonomy = 1)
}
