% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_values.R
\name{get_trait_values}
\alias{get_trait_values}
\title{Retrieve trait descriptions}
\usage{
get_trait_values()
}
\value{
A dataframe with descriptions and possible values of each trait,
specifically columns \code{trait_name}, \code{trait_description},
\code{trait_value} as well as \code{value_desc_short} and
\code{value_desc_long}.
}
\description{
\code{get_trait_values()} retrieves the descriptions and possible values of
each trait.
}
\examples{
trait_desc <- get_trait_values()
trait_desc
}
