% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_description_query.R
\name{trait_description_query}
\alias{trait_description_query}
\title{Retrieve trait descriptions}
\usage{
trait_description_query()
}
\value{
A dataframe with descriptions and possible values of each trait,
specifically columns \code{trait_name}, \code{trait_description},
\code{trait_value} as well as \code{value_desc_short} and
\code{value_desc_long}.
}
\description{
\code{trait_description_query()} retrieves the descriptions and possible values of
each trait.
}
\examples{
trait_desc <- trait_description_query()
trait_desc
}
