% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_db}
\alias{connect_db}
\title{Connect to AVONET database}
\usage{
connect_db(username, pw)
}
\arguments{
\item{username}{Username for AVONET database}

\item{pw}{Password for AVONET database}
}
\value{
A database connection
}
\description{
Connect to AVONET database
}
\examples{
\dontrun{
# This example makes use of the `keyring` package to store users' credentials
db_user <- keyring::key_list("avonet")[1, "username"]
db_password <- keyring::key_get("avonet", username = db_user)

# Connect to the AVONET database
connect_db(username = db_user, pw = db_password)

# ...do some work...

# Disconnect from the AVONET database
disconnect_db()
}
}
\seealso{
\code{\link[=disconnect_db]{disconnect_db()}}
}
