% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_db}
\alias{connect_db}
\title{Connect to AVONET database}
\usage{
connect_db(username, pw)
}
\arguments{
\item{username}{Username for AVONET database}

\item{pw}{Password for AVONET database}
}
\value{
A database connection
}
\description{
Connect to AVONET database
}
\examples{
con <- connect_db(username = "postgres", pw = "Frankfurterstr25!")
}
