% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve metadata for an AVONET query}
\usage{
get_metadata(src_cols)
}
\arguments{
\item{src_cols}{Source columns}
}
\value{
A dataframe
}
\description{
Retrieve metadata for an AVONET query
}
\examples{
db_user <- keyring::key_list("avonet")[1, "username"]
db_password <- keyring::key_get("avonet", username = db_user)
connect_db(username = db_user, pw = db_password)

species_data <- sql_query(parameter1 = "Buteo buteo", parameter2 = 1)

src_pattern  <- "(_source|src)$"
src_cols     <- species_data[, grep(src_pattern, names(species_data))]

metadata <- get_metadata(src_cols = src_cols)

metadata
}
