% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get species data}
\usage{
get_species(x, y, inferred = FALSE)
}
\arguments{
\item{x}{Latin name of species, e.g. "Buteo buteo"}

\item{y}{Taxonomy. 1 = BirdLife, 2 = eBird and 3 = BirdTree}

\item{inferred}{Include inferred traits? Defaults to FALSE. For now, only FALSE is implemented.}
}
\value{
A dataframe.
}
\description{
Get species data
}
\examples{
\dontrun{
db_user <- keyring::key_list("avonet")[1, "username"]
db_password <- keyring::key_get("avonet", username = db_user)
connect_db(username = db_user, pw = db_password)

# This works for a single species
get_species("Buteo buteo", 1)

# Or a list of several species
my_birds <- c("Haliaeetus leucocephalus", "Aptenodytes forsteri", "Cardinalis cardinalis")
get_species(my_birds, 1)
}
}
