% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get species data}
\usage{
get_species(con, x, y, inferred = FALSE)
}
\arguments{
\item{con}{Connection to the AVONET database}

\item{x}{Latin name of species, e.g. "Buteo buteo"}

\item{y}{Taxonomy. 1 = BirdLife, 2 = eBird and 3 = BirdTree}

\item{inferred}{Include inferred traits? Defaults to FALSE. For now, only FALSE is implemented.}
}
\value{
A dataframe.
}
\description{
Get species data
}
\examples{

get_species("Buteo buteo", 1)

my_birds <- c("Haliaeetus leucocephalus", "Aptenodytes forsteri", "Cardinalis cardinalis")
get_species(my_birds, 1)

}
