% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomic_info.R
\name{get_taxonomic_info}
\alias{get_taxonomic_info}
\title{Retrieve taxonomic information}
\usage{
get_taxonomic_info(con, search_term, taxonomy)
}
\arguments{
\item{con}{Connection to the AVONET database}

\item{search_term}{Name of a genus, family or order}

\item{taxonomy}{Choose which taxonomy your results are displayed in. 1 = BirdLife, 2 = eBird and 3 = BirdTree}
}
\value{
A dataframe
}
\description{
Retrieve taxonomic information
}
\examples{
con <- connect_db(username = "postgres", pw = "Frankfurterstr25!")
get_taxonomic_info(con, search_term = "Buteo", taxonomy = 1)

get_taxonomic_info(con, search_term = "passeriformes", taxonomy = 1)
}
