% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_groups.R
\name{get_trait_groups}
\alias{get_trait_groups}
\title{Discover available trait groups}
\usage{
get_trait_groups()
}
\value{
A character vector of trait groups, e.g. "eco", "morpho", "geo"
}
\description{
\code{get_trait_groups()} returns a character vector of the groups of traits that
are available in the AVONET dataset.
}
\examples{
traitGroups <- get_trait_groups()
traitGroups
}
