% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_groups.R
\name{get_trait_groups}
\alias{get_trait_groups}
\title{Discover available trait groups}
\usage{
get_trait_groups()
}
\value{
A character vector of trait groups: "eco", "social", "morpho" and
"geo". Further trait groups, e.g. "reproductive" and "demo", will be
implemented soon.
}
\description{
\code{get_trait_groups()} returns a character vector of the groups of traits that
are available in the AVONET dataset.
}
\examples{
\dontrun{
traitGroups <- get_trait_groups()
traitGroups
}
}
\seealso{
\code{\link[=get_trait_list]{get_trait_list()}}, which takes one of these names as its \code{group}
argument.
}
