% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_list.R
\name{get_trait_list}
\alias{get_trait_list}
\title{See which traits belong to a trait group}
\usage{
get_trait_list(group = NULL)
}
\arguments{
\item{group}{A string. Currently accepts one of the following values: "eco",
"morpho" or "geo". If NULL, then \code{get_trait_list()} returns all traits from every group.}
}
\value{
A tibble of traits
}
\description{
\code{get_trait_list()} returns a tibble of the traits that are available in a specific
trait group. In addition to the traits, it provides a description, possible values,
sources and the resolution of the trait (species vs. specimen).
}
\examples{
\dontrun{
eco_traits <- get_trait_list(group = "eco")
eco_traits
}
}
