% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.R
\name{sql_query}
\alias{sql_query}
\title{Create SQL query of AVONET database}
\usage{
sql_query(parameter1, parameter2)
}
\arguments{
\item{parameter1}{Latin name of species}

\item{parameter2}{Taxonomy}
}
\value{
A dataframe.
}
\description{
Create SQL query of AVONET database
}
\examples{
\dontrun{
db_user <- keyring::key_list("avonet")[1, "username"]
db_password <- keyring::key_get("avonet", username = db_user)
connect_db(username = db_user, pw = db_password)

parameter1 <- "Buteo buteo"
parameter2 <- 1
result_df <- sql_query(parameter1, parameter2)
result_df
}
}
