% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.R
\name{sql_query}
\alias{sql_query}
\title{Create SQL query of AVONET database}
\usage{
sql_query(con, parameter1, parameter2)
}
\arguments{
\item{con}{Connection to the AVONET database}

\item{parameter1}{Latin name of species}

\item{parameter2}{Taxonomy}
}
\value{
A dataframe.
}
\description{
Create SQL query of AVONET database
}
\examples{
con <- connect(username = "postgres", pw = "Frankfurterstr25!")
parameter1 <- "Buteo buteo"
parameter2 <- 1
result_df <- sql_query(con, parameter1, parameter2)
result_df
}
